requirements/dev.txt
-r base.txt

mocker==1.1.1
pluggy==1.0.0
pre-commit==3.3.2
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
PyYAML==6.0
